fix: move GetReplicationDetailsByReplicationId method
